+++ typescript/ui/src/core/stores/HistoryStore/HistoryStoreContextProvider.tsx
@@ -1,77 +1,97 @@
-import { PropsWithChildren, useState } from "react";
-import { createStore } from "zustand";
+import { PropsWithChildren, useMemo, useState } from "react";
+import { createStore, useStore } from "zustand";
 import { createHistoryStore } from "./historyStore";
 import { createJSONStorage, persist } from "zustand/middleware";
 import { IndexedDBStorage } from "./indexedDBStorage";
 import { HistoryStoreContext } from "./HistoryStoreContext";
-import { transform } from "lodash";
 import { Conversation } from "../../../types/history";
+import InitializationScreen from "../../components/InitializationScreen";
+
+export const HISTORY_STORE_KEY_BASE = "ragbits-history-store";
 
 interface HistoryStoreContextProviderProps {
   shouldStoreHistory: boolean;
 }
 
-export function HistoryStoreContextProvider({
+function initializeStore(shouldStoreHistory: boolean, storeKey: string) {
+  if (shouldStoreHistory) {
+    return createStore(
+      persist(createHistoryStore, {
+        name: storeKey,
+        partialize: (value) => ({
+          conversations: value.conversations,
+        }),
+        onRehydrateStorage: (state) => {
+          // We have to wait for the hydration to finish to avoid any races that may result in saving of the invalid
+          // state to the storage (e.g. clearing all history)
+          return () => state._internal._setHasHydrated(true);
+        },
+        merge: (persistedState, currentState) => {
+          const persistedConversations =
+            (persistedState as Record<string, unknown>)?.conversations ?? {};
+          const { conversations, currentConversation } = currentState;
+          const fixedConversations = Object.values(
+            persistedConversations,
+          ).reduce((acc, c: Conversation) => {
+            if (c.conversationId === null) {
+              return acc;
+            }
+
+            acc[c.conversationId] = {
+              ...c,
+              isLoading: false,
+              abortController: null,
+            };
+            return acc;
+          }, {});
+
+          return {
+            ...currentState,
+            currentConversation: currentConversation,
+            conversations: { ...fixedConversations, ...conversations },
+          };
+        },
+        storage: createJSONStorage(() => IndexedDBStorage),
+      }),
+    );
+  }
+
+  // Manually set _hasHydrated when we don't use any storage
+  const store = createStore(createHistoryStore);
+  store.getState()._internal._setHasHydrated(true);
+
+  return store;
+}
+
+export default function HistoryStoreContextProvider({
   children,
   shouldStoreHistory,
 }: PropsWithChildren<HistoryStoreContextProviderProps>) {
-  const [store] = useState(() =>
-    shouldStoreHistory
-      ? createStore(
-          persist(
-            createHistoryStore,
+  const [storeKey, _setStoreKey] = useState(HISTORY_STORE_KEY_BASE);
+  const store = useMemo(
+    () => initializeStore(shouldStoreHistory, storeKey),
+    [shouldStoreHistory, storeKey],
+  );
+  const hasHydrated = useStore(store, (s) => s._internal._hasHydrated);
 
-            {
-              name: "ragbits-history-store",
-              partialize: (value) => ({
-                conversations: value.conversations,
-              }),
-              merge: (persistedState, currentState) => {
-                const { conversations, currentConversation } = currentState;
-                const finalState = {
-                  ...currentState,
-                  ...(persistedState ?? {}),
-                };
+  const initializeUserStore = (userId: string) => {
+    _setStoreKey(`${HISTORY_STORE_KEY_BASE}-${userId}`);
+  };
 
-                // When loading the state, we need to reset the `isLoading` and `abortController` properties,
-                // since any information about the current stream is lost when the app is closed or reloaded.
-                finalState.conversations = transform<
-                  Conversation,
-                  Record<string, Conversation>
-                >(
-                  finalState.conversations,
-                  (res, c) => {
-                    // Ignore old conversations with `null` as key
-                    if (c.conversationId === null) {
-                      return;
-                    }
-
-                    const newKey = c.conversationId;
-                    const newValue = {
-                      ...c,
-                      isLoading: false,
-                      abortController: null,
-                    };
-                    res[newKey] = newValue;
-                  },
-                  {},
-                );
-                // This ensures that we always start with empty conversation
-                finalState.conversations[currentConversation] =
-                  conversations[currentConversation];
-                finalState.currentConversation = currentConversation;
-
-                return finalState;
-              },
-              storage: createJSONStorage(() => IndexedDBStorage),
-            },
-          ),
-        )
-      : createStore(createHistoryStore),
+  const value = useMemo(
+    () => ({
+      store,
+      initializeUserStore,
+    }),
+    [store],
   );
 
+  if (shouldStoreHistory && !hasHydrated) {
+    return <InitializationScreen />;
+  }
+
   return (
-    <HistoryStoreContext.Provider value={store}>
+    <HistoryStoreContext.Provider value={value}>
       {children}
     </HistoryStoreContext.Provider>
   );