+++ packages/ragbits-chat/CHANGELOG.md
@@ -1,6 +1,16 @@
 # CHANGELOG
 
 ## Unreleased
+- Remove redundant test for `/api/config` endpoint (#795)
+- Fix bug causing infinite initialization screen (#793)
+- Fix bug that caused messages to be sent when changing chat settings; simplify and harden history logic (#791)
+- Add `clear_message` event type allowing to reset whole message (#789)
+- Add usage component to UI with backend support (#786)
+- Add authentication handling in the UI (#763)
+- Add backend authentication (#761)
+
+- Autogenerate typescript types based on backend typing (#727)
+- Add ability to customize favicon and page title (#766)
 
 - Autogenerate typescript types based on backend typing (#727)
 - Add ability to customize favicon and page title (#766)